Postman2

+++ src/main/java/lt/bit/products/store/controller/ProductController.java
@@ -2,12 +2,15 @@ package lt.bit.products.store.controller;
 
 import java.util.List;
 import lt.bit.products.store.model.Product;
+import lt.bit.products.store.model.ProductRequest;
 import lt.bit.products.store.service.ProductService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
@@ -21,6 +24,12 @@ class ProductController {
 
   ProductController(ProductService service) {
     this.service = service;
+  }
+
+  @ResponseStatus(HttpStatus.CREATED)
+  @PostMapping
+  Product createProduct(@RequestBody ProductRequest productRequest) {
+    return null;//service.saveProduct();
   }
 
   @GetMapping
@@ -37,9 +46,13 @@ class ProductController {
     return ResponseEntity.ok(product);
   }
 
-  @ResponseStatus(HttpStatus.NO_CONTENT)
   @DeleteMapping("/{id}")
-  void deleteProduct(@PathVariable("id") Integer productId) {
+  ResponseEntity<Void> deleteProduct(@PathVariable("id") Integer productId) {
+    Product product = service.findProduct(productId);
+    if (product == null) {
+      return ResponseEntity.notFound().build();
+    }
     service.deleteProduct(productId);
+    return ResponseEntity.noContent().build();
   }
 }
